Pass between scenes
An entery scene was created.
A transition was added between the entery scene and the training scene.
With the FadeController.cs script, the screen goes dark for 1 second during the transition between levels and then moves to a new scene.
With the ScenesPassController.cs script, the player's movement is restricted while passing through the doors and the player waits until the transition continues.